+++ tests/core/gui/test_file_history.py
@@ -1,9 +1,7 @@
 # Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
 
 import pytest
-from PyQt5 import QtCore, QtWidgets
-
-from tests.common import customize_fixtures
+from PyQt5 import QtWidgets
 
 
 @pytest.fixture
@@ -11,92 +9,55 @@ def catch_file_history_widget(widget_catcher_factory):
     return widget_catcher_factory("parsec.core.gui.file_history_widget.FileHistoryWidget")
 
 
-async def create_workspace(aqtbot, logged_gui, monkeypatch):
-    w_w = await logged_gui.test_switch_to_workspaces_widget()
-    add_button = w_w.button_add_workspace
-    assert add_button is not None
-
-    monkeypatch.setattr(
-        "parsec.core.gui.workspaces_widget.get_text_input", lambda *args, **kwargs: ("Workspace")
-    )
-    async with aqtbot.wait_signals(
-        [w_w.create_success, w_w.list_success, w_w.mountpoint_started], timeout=2000
-    ):
-        await aqtbot.mouse_click(add_button, QtCore.Qt.LeftButton)
-
-    def _workspace_button_ready():
-        assert w_w.layout_workspaces.count() == 1
-        wk_button = w_w.layout_workspaces.itemAt(0).widget()
-        assert not isinstance(wk_button, QtWidgets.QLabel)
-
-    await aqtbot.wait_until(_workspace_button_ready, timeout=2000)
-    wk_button = w_w.layout_workspaces.itemAt(0).widget()
-    assert wk_button.name == "Workspace"
-
-    async with aqtbot.wait_signal(w_w.load_workspace_clicked):
-        await aqtbot.mouse_click(wk_button, QtCore.Qt.LeftButton)
-
-    return w_w
-
-
-async def create_directories(logged_gui, aqtbot, monkeypatch, dir_names):
-    central_widget = logged_gui.test_get_central_widget()
-    assert central_widget is not None
-
-    f_w = logged_gui.test_get_files_widget()
-    assert f_w is not None
-
-    add_button = f_w.button_create_folder
-
-    for dir_name in dir_names:
-        monkeypatch.setattr(
-            "parsec.core.gui.files_widget.get_text_input", lambda *args, **kwargs: (dir_name)
-        )
-        async with aqtbot.wait_signal(f_w.folder_create_success):
-            await aqtbot.mouse_click(add_button, QtCore.Qt.LeftButton)
-
-    async with aqtbot.wait_signals([f_w.folder_stat_success, f_w.fs_synced_qt], timeout=3000):
-        pass
-
-    f_w.table_files.rowCount() == 1
-
-    def _folder_synced():
-        item = f_w.table_files.item(1, 0)
-        assert item.is_synced
-
-    await aqtbot.wait_until(_folder_synced, timeout=3000)
-    return f_w
-
-
 @pytest.mark.gui
 @pytest.mark.trio
-@pytest.mark.flaky(reruns=1)
-@customize_fixtures(logged_gui_as_admin=True)
 async def test_file_history(
-    aqtbot,
-    running_backend,
-    logged_gui,
-    monkeypatch,
-    autoclose_dialog,
-    qt_thread_gateway,
-    catch_file_history_widget,
-    alice,
+    aqtbot, running_backend, logged_gui, monkeypatch, autoclose_dialog, catch_file_history_widget
 ):
-    await create_workspace(aqtbot, logged_gui, monkeypatch)
-    f_w = await create_directories(logged_gui, aqtbot, monkeypatch, ["dir1"])
+    core = logged_gui.test_get_core()
+    wid = await core.user_fs.workspace_create("wksp1")
+    wfs = core.user_fs.get_workspace(wid)
 
+    w_w = logged_gui.test_get_workspaces_widget()
+
+    def _workspace_available():
+        assert w_w.layout_workspaces.count() == 1
+
+    await aqtbot.wait_until(_workspace_available)
+
+    f_w = await logged_gui.test_switch_to_files_widget("wksp1")
+
+    # Add an entry to the workspace
+    await wfs.touch("/file.txt")
+    await wfs.sync()
+    await wfs.write_bytes("/file.txt", data=b"v2")
+    await wfs.sync()
+
+    def _entry_available():
+        assert f_w.table_files.rowCount() == 2
+
+    await aqtbot.wait_until(_entry_available)
+
+    # First select the entry...
     await aqtbot.run(
         f_w.table_files.setRangeSelected, QtWidgets.QTableWidgetSelectionRange(1, 0, 1, 0), True
     )
 
-    await qt_thread_gateway.send_action(f_w.show_history)
+    def _entry_selected():
+        assert f_w.table_files.selectedItems()
+
+    await aqtbot.wait_until(_entry_selected)
+
+    # ...then ask for history
+    await aqtbot.run(f_w.table_files.show_history_clicked.emit)
+
     hf_w = await catch_file_history_widget()
 
-    def _history_filled():
+    def _history_displayed():
+        assert hf_w.isVisible()
         assert hf_w.layout_history.count() == 1
-        assert hf_w.layout_history.itemAt(0).widget()
+        hb2_w = hf_w.layout_history.itemAt(0).widget()
+        assert hb2_w.label_user.text() == "Boby McBobFace"
+        assert hb2_w.label_version.text() == "2"
 
-    await aqtbot.wait_until(_history_filled, timeout=5000)
-    history_button = hf_w.layout_history.itemAt(0).widget()
-    assert history_button.label_user.text() == alice.human_handle.label
-    assert history_button.label_version.text() == "1"
+    await aqtbot.wait_until(_history_displayed)