+++ tests/core/gui/test_workspaces.py
@@ -1,118 +1,106 @@
 # Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
 
-from parsec.core.core_events import CoreEvent
 import pytest
 from PyQt5 import QtCore
-
-from parsec.core.local_device import save_device_with_password
-from parsec.core.fs import FSWorkspaceNoReadAccess
-from unittest.mock import ANY
-from parsec.api.data import WorkspaceEntry
 from uuid import UUID
 import pendulum
+from unittest.mock import ANY
 
-
-@pytest.fixture
-async def logged_gui(aqtbot, gui_factory, autoclose_dialog, core_config, alice, bob):
-    save_device_with_password(core_config.config_dir, alice, "P@ssw0rd")
-
-    gui = await gui_factory()
-    lw = gui.test_get_login_widget()
-    tabw = gui.test_get_tab()
-
-    assert lw is not None
-
-    await aqtbot.key_clicks(lw.line_edit_password, "P@ssw0rd")
-
-    async with aqtbot.wait_signals([lw.login_with_password_clicked, tabw.logged_in]):
-        await aqtbot.mouse_click(lw.button_login, QtCore.Qt.LeftButton)
-
-    central_widget = gui.test_get_central_widget()
-    assert central_widget is not None
-
-    save_device_with_password(core_config.config_dir, bob, "P@ssw0rd")
-
-    yield gui
+from parsec.api.data import WorkspaceEntry
+from parsec.core.core_events import CoreEvent
+from parsec.core.fs import FSWorkspaceNoReadAccess
+from parsec.core.gui.workspace_button import WorkspaceButton
 
 
 @pytest.mark.gui
 @pytest.mark.trio
+@pytest.mark.flaky(reruns=1)
 @pytest.mark.parametrize("invalid_name", (False, True))
-async def test_add_workspace(aqtbot, running_backend, logged_gui, monkeypatch, invalid_name):
-    w_w = logged_gui.test_get_workspaces_widget()
+async def test_add_workspace(
+    aqtbot, running_backend, logged_gui, monkeypatch, autoclose_dialog, invalid_name
+):
+    w_w = await logged_gui.test_switch_to_workspaces_widget()
 
-    assert w_w is not None
-    async with aqtbot.wait_signal(w_w.list_success):
-        pass
+    # Make sure there is no workspaces to display
     assert w_w.layout_workspaces.count() == 1
     assert w_w.layout_workspaces.itemAt(0).widget().text() == "No workspace has been created yet."
 
-    add_button = w_w.button_add_workspace
-    add_button is not None
-
+    # Add (or try to) a new workspace
     workspace_name = ".." if invalid_name else "Workspace1"
     monkeypatch.setattr(
         "parsec.core.gui.workspaces_widget.get_text_input", lambda *args, **kwargs: (workspace_name)
     )
+    await aqtbot.mouse_click(w_w.button_add_workspace, QtCore.Qt.LeftButton)
 
-    if invalid_name:
-        async with aqtbot.wait_signals([w_w.create_error, w_w.list_success], timeout=2000):
-            await aqtbot.mouse_click(add_button, QtCore.Qt.LeftButton)
-
+    def _outcome_occured():
         assert w_w.layout_workspaces.count() == 1
-        assert (
-            w_w.layout_workspaces.itemAt(0).widget().text() == "No workspace has been created yet."
-        )
+        if invalid_name:
+            assert (
+                w_w.layout_workspaces.itemAt(0).widget().text()
+                == "No workspace has been created yet."
+            )
+            assert autoclose_dialog.dialogs == [
+                (
+                    "Error",
+                    "Could not create the workspace. This name is not a valid workspace name.",
+                )
+            ]
+        else:
+            wk_button = w_w.layout_workspaces.itemAt(0).widget()
+            assert isinstance(wk_button, WorkspaceButton)
+            assert wk_button.name == "Workspace1"
+            assert not autoclose_dialog.dialogs
 
-    else:
-        async with aqtbot.wait_signals([w_w.create_success, w_w.list_success], timeout=2000):
-            await aqtbot.mouse_click(add_button, QtCore.Qt.LeftButton)
-
-        assert w_w.layout_workspaces.count() == 1
-        wk_button = w_w.layout_workspaces.itemAt(0).widget()
-        assert wk_button.name == "Workspace1"
+    await aqtbot.wait_until(_outcome_occured)
 
 
 @pytest.mark.gui
 @pytest.mark.trio
+@pytest.mark.flaky(reruns=1)
 @pytest.mark.parametrize("invalid_name", (False, True))
-async def test_rename_workspace(aqtbot, running_backend, logged_gui, monkeypatch, invalid_name):
-    w_w = logged_gui.test_get_workspaces_widget()
+async def test_rename_workspace(
+    aqtbot, running_backend, logged_gui, monkeypatch, autoclose_dialog, invalid_name
+):
+    w_w = await logged_gui.test_switch_to_workspaces_widget()
 
-    assert w_w is not None
-    async with aqtbot.wait_signal(w_w.list_success):
-        pass
-    assert w_w.layout_workspaces.count() == 1
-    assert w_w.layout_workspaces.itemAt(0).widget().text() == "No workspace has been created yet."
+    # Create a workspace and make sure the workspace is displayed
+    core = logged_gui.test_get_core()
+    await core.user_fs.workspace_create("Workspace1")
 
-    add_button = w_w.button_add_workspace
-    assert add_button is not None
+    def _workspace_displayed():
+        assert w_w.layout_workspaces.count() == 1
+        wk_button = w_w.layout_workspaces.itemAt(0).widget()
+        assert isinstance(wk_button, WorkspaceButton)
+        assert wk_button.name == "Workspace1"
 
-    monkeypatch.setattr(
-        "parsec.core.gui.workspaces_widget.get_text_input", lambda *args, **kwargs: ("Workspace1")
-    )
-
-    async with aqtbot.wait_signals([w_w.create_success, w_w.list_success], timeout=2000):
-        await aqtbot.mouse_click(add_button, QtCore.Qt.LeftButton)
-
-    assert w_w.layout_workspaces.count() == 1
+    await aqtbot.wait_until(_workspace_displayed)
     wk_button = w_w.layout_workspaces.itemAt(0).widget()
-    assert wk_button.name == "Workspace1"
 
+    # Now do the rename
     workspace_name = ".." if invalid_name else "Workspace1_Renamed"
     monkeypatch.setattr(
         "parsec.core.gui.workspaces_widget.get_text_input", lambda *args, **kwargs: (workspace_name)
     )
+    await aqtbot.mouse_click(wk_button.button_rename, QtCore.Qt.LeftButton)
 
-    if invalid_name:
-        async with aqtbot.wait_signal(w_w.rename_error):
-            await aqtbot.mouse_click(wk_button.button_rename, QtCore.Qt.LeftButton)
-        assert wk_button.name == "Workspace1"
+    def _outcome_occured():
+        assert w_w.layout_workspaces.count() == 1
+        new_wk_button = w_w.layout_workspaces.itemAt(0).widget()
+        assert isinstance(new_wk_button, WorkspaceButton)
+        assert new_wk_button.workspace_fs is wk_button.workspace_fs
+        if invalid_name:
+            assert wk_button.name == "Workspace1"
+            assert autoclose_dialog.dialogs == [
+                (
+                    "Error",
+                    "Could not rename the workspace. This name is not a valid workspace name.",
+                )
+            ]
+        else:
+            assert wk_button.name == "Workspace1_Renamed"
+            assert not autoclose_dialog.dialogs
 
-    else:
-        async with aqtbot.wait_signal(w_w.rename_success):
-            await aqtbot.mouse_click(wk_button.button_rename, QtCore.Qt.LeftButton)
-        assert wk_button.name == "Workspace1_Renamed"
+    await aqtbot.wait_until(_outcome_occured)
 
 
 @pytest.mark.skip("No notification center at the moment")
@@ -148,10 +136,11 @@ async def test_mountpoint_remote_error_event(aqtbot, running_backend, logged_gui
     )
 
 
+@pytest.mark.skip("Should be reworked")
 @pytest.mark.gui
 @pytest.mark.trio
-async def test_event_bus_internal_connection(aqtbot, running_backend, logged_gui, alice):
-    w_w = logged_gui.test_get_workspaces_widget()
+async def test_event_bus_internal_connection(aqtbot, running_backend, logged_gui, autoclose_dialog):
+    w_w = await logged_gui.test_switch_to_workspaces_widget()
     uuid = UUID("1bc1e17b-157a-462f-86f2-7f64657ba16a")
     w_entry = WorkspaceEntry(
         name="w",
@@ -162,10 +151,6 @@ async def test_event_bus_internal_connection(aqtbot, running_backend, logged_gui
         role_cached_on=ANY,
         role=None,
     )
-
-    assert w_w is not None
-    async with aqtbot.wait_signal(w_w.list_success):
-        pass
 
     async with aqtbot.wait_signal(w_w.fs_synced_qt):
         w_w.event_bus.send(CoreEvent.FS_ENTRY_SYNCED, workspace_id=None, id=uuid)
@@ -190,6 +175,7 @@ async def test_event_bus_internal_connection(aqtbot, running_backend, logged_gui
             timestamp=pendulum.now(),
         )
 
+    assert not autoclose_dialog.dialogs
     async with aqtbot.wait_signal(w_w.mountpoint_stopped):
         w_w.event_bus.send(
             CoreEvent.MOUNTPOINT_STOPPED,
@@ -197,3 +183,9 @@ async def test_event_bus_internal_connection(aqtbot, running_backend, logged_gui
             workspace_id=uuid,
             timestamp=pendulum.now(),
         )
+    assert autoclose_dialog.dialogs == [
+        (
+            "Error",
+            "Your permissions on this workspace have been revoked. You no longer have access to theses files.",
+        )
+    ]